// src/models/schemas/Bookmarks.schema.ts
import { ObjectId } from 'mongodb'

interface BookmarkType {
  _id?: ObjectId
  user_id: ObjectId
  tweet_id: ObjectId
  creater_at?: Date
}
export default class Bookmark {
  _id: ObjectId
  user_id: ObjectId
  tweet_id: ObjectId
  creater_at?: Date
  constructor({ _id, user_id, tweet_id, creater_at }: BookmarkType) {
    this._id = _id || new ObjectId()
    this.user_id = user_id
    this.tweet_id = tweet_id
    this.creater_at = creater_at || new Date()
  }
}